requirements.txt
# Core Dependencies
supabase==2.3.0
python-dotenv==1.0.0

# PDF Processing
PyMuPDF==1.23.8
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==10.1.0

# AI/ML
google-generativeai==0.3.2

# Utilities
python-dateutil==2.8.2
pydantic==2.5.3

# Development (Optional)
pytest==7.4.3
black==23.12.1
flake8==7.0.0

fastapi
uvicorn
python-multipart

